backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
